+++ internal/jbig2/segments/symbol-dictionary_test.go
@@ -47,8 +47,8 @@ func TestSymbolDictionaryDecode(t *testing.T) {
 
 		assert.True(t, s.isHuffmanEncoded)
 		assert.False(t, s.useRefinementAggregation)
-		assert.Equal(t, uint32(1), s.numberOfExportedSymbols)
-		assert.Equal(t, uint32(1), s.numberOfNewSymbols)
+		assert.Equal(t, 1, s.numberOfExportedSymbols)
+		assert.Equal(t, 1, s.numberOfNewSymbols)
 
 		bm, err := s.GetDictionary()
 		require.NoError(t, err)
@@ -88,8 +88,8 @@ func TestSymbolDictionaryDecode(t *testing.T) {
 
 		assert.True(t, s.isHuffmanEncoded)
 		assert.False(t, s.useRefinementAggregation)
-		assert.Equal(t, uint32(2), s.numberOfExportedSymbols)
-		assert.Equal(t, uint32(2), s.numberOfNewSymbols)
+		assert.Equal(t, 2, s.numberOfExportedSymbols)
+		assert.Equal(t, 2, s.numberOfNewSymbols)
 
 		bm, err := s.GetDictionary()
 		require.NoError(t, err)
@@ -136,8 +136,8 @@ func TestSymbolDictionaryDecode(t *testing.T) {
 		assert.Equal(t, int8(2), s.sdATX[0])
 		assert.Equal(t, int8(-1), s.sdATY[0])
 
-		assert.Equal(t, uint32(2), s.numberOfExportedSymbols)
-		assert.Equal(t, uint32(2), s.numberOfNewSymbols)
+		assert.Equal(t, 2, s.numberOfExportedSymbols)
+		assert.Equal(t, 2, s.numberOfNewSymbols)
 
 		bm, err := s.GetDictionary()
 		require.NoError(t, err)
@@ -203,8 +203,8 @@ func TestSymbolDictionaryDecode(t *testing.T) {
 			assert.Equal(t, s.sdATY[0], int8(-1))
 		}
 
-		assert.Equal(t, uint32(1), s.numberOfExportedSymbols)
-		assert.Equal(t, uint32(1), s.numberOfNewSymbols)
+		assert.Equal(t, 1, s.numberOfExportedSymbols)
+		assert.Equal(t, 1, s.numberOfNewSymbols)
 
 		bm, err := s.GetDictionary()
 		require.NoError(t, err)
@@ -229,8 +229,8 @@ func isLetterP(t *testing.T, b *bitmap.Bitmap) {
 		return
 	}
 
-	for y := int32(0); y < b.Height; y++ {
-		for x := int32(0); x < b.Width; x++ {
+	for y := 0; y < b.Height; y++ {
+		for x := 0; x < b.Width; x++ {
 			pix := b.GetPixel(x, y)
 
 			switch y {
@@ -263,8 +263,8 @@ func isLetterC(t *testing.T, b *bitmap.Bitmap) {
 		return
 	}
 
-	for y := int32(0); y < b.Height; y++ {
-		for x := int32(0); x < b.Width; x++ {
+	for y := 0; y < b.Height; y++ {
+		for x := 0; x < b.Width; x++ {
 			pix := b.GetPixel(x, y)
 
 			switch y {
@@ -297,8 +297,8 @@ func isLetterA(t *testing.T, b *bitmap.Bitmap) {
 		return
 	}
 
-	for y := int32(0); y < b.Height; y++ {
-		for x := int32(0); x < b.Width; x++ {
+	for y := 0; y < b.Height; y++ {
+		for x := 0; x < b.Width; x++ {
 			pix := b.GetPixel(x, y)
 
 			switch y {
